backend/requirements.txt
Flask
flask-cors
flask-socketio
eventlet
requests
tornado
python-socketio